added a software design document, research document, made progress on the...

See merge request I450502/quizwebapp!4

+++ QuizWebapp/src/main/java/IndividualProject/persistence/entity/User.java
@@ -10,6 +10,6 @@ public class User {
     private String name;
     private String role;
     private long id;
-    private long totalScore;
+    private Integer totalScore;
 
 }
